+++ lib/yard-sorbet/handlers/sig_handler.rb
@@ -64,15 +64,13 @@ module YARDSorbet
         attrs = namespace.attributes[scope][names[0]]
         return false if attrs.nil? || attrs.empty?
 
-        document_attr_methods(attrs.values.compact)
+        document_attrs(attrs.values.compact)
         attr_node.docstring = nil
         true
       end
 
       sig { params(method_objects: T::Array[YARD::CodeObjects::MethodObject]).void }
-      def document_attr_methods(method_objects)
-        method_objects.each { parse_node(_1, _1.docstring, include_params: false) }
-      end
+      def document_attrs(method_objects) =  method_objects.each { parse_node(_1, _1.docstring, include_params: false) }
 
       sig { params(attach_to: Documentable, docstring: T.nilable(String), include_params: T::Boolean).void }
       def parse_node(attach_to, docstring, include_params: true)