+++ lib/yard-sorbet/handlers/struct_prop_handler.rb
@@ -35,15 +35,11 @@ module YARDSorbet
       end
 
       sig { returns(T::Boolean) }
-      def immutable?
-        statement.method_name(true) == :const || kw_arg('immutable:') == 'true'
-      end
+      def immutable? = statement.method_name(true) == :const || kw_arg('immutable:') == 'true'
 
       # @return the value passed to the keyword argument, or nil
       sig { params(kwd: String).returns(T.nilable(String)) }
-      def kw_arg(kwd)
-        params[2]&.find { _1[0].source == kwd }&.[](1)&.source
-      end
+      def kw_arg(kwd) = params[2]&.find { _1[0].source == kwd }&.[](1)&.source
 
       sig { params(name: String).returns(TStructProp) }
       def make_prop(name)
@@ -57,9 +53,7 @@ module YARDSorbet
       end
 
       sig { returns(T::Array[YARD::Parser::Ruby::AstNode]) }
-      def params
-        @params ||= T.let(statement.parameters(false), T.nilable(T::Array[YARD::Parser::Ruby::AstNode]))
-      end
+      def params = @params ||= T.let(statement.parameters(false), T.nilable(T::Array[YARD::Parser::Ruby::AstNode]))
 
       # Register the field explicitly as an attribute.
       sig { params(object: YARD::CodeObjects::MethodObject, name: String).void }