+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Changes
 
-* [#282](https://github.com/dduugg/yard-sorbet/issues/282) Increase minimum required ruby version to `3.0`
+* [#311](https://github.com/dduugg/yard-sorbet/issues/311) Increase minimum required ruby version to `3.1`
 
 ## 0.8.1 (2023-04-03)